+++ components/src/core/drawer/drawer-subheader/drawer-subheader.component.ts
@@ -7,7 +7,10 @@ import { StateListener } from '../state-listener.component';
     changeDetection: ChangeDetectionStrategy.OnPush,
     encapsulation: ViewEncapsulation.None,
     template: `
-        <div class="pxb-drawer-subheader-content" [class.pxb-drawer-subheader-closed]="!isOpen()">
+        <div
+            class="pxb-drawer-subheader-content"
+            [style.visibility]="hideContentOnCollapse ? (isOpen() ? 'visible' : 'hidden') : 'visible'"
+        >
             <ng-content></ng-content>
         </div>
         <mat-divider *ngIf="divider"></mat-divider>
@@ -19,6 +22,7 @@ import { StateListener } from '../state-listener.component';
 })
 export class DrawerSubheaderComponent extends StateListener {
     @Input() divider = true;
+    @Input() hideContentOnCollapse = true;
 
     constructor(drawerService: DrawerService, changeDetectorRef: ChangeDetectorRef) {
         super(drawerService, changeDetectorRef);