+++ components/src/core/user-menu/public-api.ts
@@ -1,2 +1,3 @@
 export * from './user-menu.component';
 export * from './user-menu.module';
+export * from './user-menu-avatar.component';